1/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
openai==1.3.7
chromadb==0.4.18
langchain==0.0.350
langchain-openai==0.0.2
langchain-chroma==0.2.5
langchain-huggingface==0.3.1
dashscope==1.23.9
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.4
numpy==1.25.2
pillow==10.1.0
python-docx==1.1.0
jinja2==3.1.2
qiniu==7.13.0
requests==2.31.0 